+++ src/utils/auth.ts
@@ -1,4 +1,9 @@
 const isLogin = () =>
+  !!(
+    localStorage.getItem('accessToken') || sessionStorage.getItem('accessToken')
+  );
+
+const getAccessToken = () =>
   localStorage.getItem('accessToken') || sessionStorage.getItem('accessToken');
 
-export { isLogin };
+export { isLogin, getAccessToken };
