+++ src/utils/dom.ts
@@ -1,0 +1,6 @@
+const createInputValueGetter =
+  (formElement: HTMLFormControlsCollection) => (id: string) => {
+    return (formElement.namedItem(id) as HTMLInputElement).value;
+  };
+
+export { createInputValueGetter };
